requirements.txt
# TrustGraph Engine Dependencies - Python 3.12 Compatible

# Core Web Framework (Essential)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AWS SDK and Services
boto3>=1.34.0
botocore>=1.34.0
aws-lambda-powertools>=2.29.0

# Voice Processing and AI (Simplified for demo)
pydub>=0.25.1
speechrecognition>=3.10.0
gtts>=2.4.0
requests>=2.31.0

# Blockchain and Cryptography
cryptography>=41.0.0
ecdsa>=0.18.0

# Machine Learning (Core only)
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Authentication and Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Utilities and Helpers
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic-settings>=2.1.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Monitoring and Logging
structlog>=23.2.0

# Optional: Advanced ML (install separately if needed)
# torch>=2.1.0  # Uncomment if you need PyTorch
# transformers>=4.30.0  # Uncomment for advanced NLP